requirements.txt
scipy
matplotlib
pillow
imageio
